requirements.txt
# For Python
setuptools>=60.1.0

# For FastAPI Application
fastapi>=0.81.0

# Other useful dependencies
# used in the Flask web app
aiohttp>=3.8.1

# For integrity checks
google-crc32c>=1.3.0

# For Google Cloud Platform APIs
google-api-python-client>=2.51.0

# For Google Cloud Platform KMS
cryptography>=37.0.2
google-cloud-kms>=2.11.2

# Other Google Cloud Platform APIs
# used in the Flask web app
google-cloud-secret-manager>=2.11.1
google-cloud-logging>=3.1.2